pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django>=5.1.6",
    "django-allauth[socialaccount]>=65.4.1",
    "django-browser-reload>=1.18.0",
    "django-debug-toolbar>=5.0.1",
    "django-environ>=0.12.0",
    "django-grappelli>=4.0.1",
    "django-template-partials>=24.4",
    "markdown>=3.7",
    "markdown-headdown>=0.1.3",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.0.1",
]
